refactor: логика виджетов перенесена на страницы, добавлена единая точка входа в компонентах
