+++ src/pages/404/404.ts
@@ -1,4 +1,4 @@
-import { renderLayoutCentered } from "shared/ui/layouts/centered/centered";
+import { renderLayoutCentered } from "shared/ui/layouts/centered";
 import { renderCreator } from "shared/utils/utils";
 
 import source from "./404.hbs";
