+++ src/shared/ui/layouts/centered/index.ts
@@ -1,0 +1,1 @@
+export * from "./centered";
