+++ src/shared/ui/list-v1/index.ts
@@ -1,0 +1,1 @@
+export * from "./list-v1";
